Allow uninstaller to run in silent mode
The /SD parameter with a value will be used to specify the option that will be used when the installer is silent.
The default options in the message boxes is to uninstall devkitPro.

How to run the silent mode with NSIS:
powershell -command start-process C:\devkitPro\uninst.exe -ArgumentList '/S' -Wait